+++ packages/react-query/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @richie-rpc/react-query
 
+## 2.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @richie-rpc/client@2.0.1
+
 ## 2.0.0
 
 ### Major Changes